add initial timestamp and tracking to desktop potentially, move analytics to core
